nvidiadockerstats/NotionDocker/requirements.txt
notion_client==2.1.0
python-dotenv==1.0.1
Requests==2.32.3
setuptools==58.1.0
pprintpp==0.4.0
regex==2023.10.3
prettytable==3.11.0
discord.py==2.3.2